requirements.txt
# Core Framework
streamlit==1.28.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Google Cloud
google-cloud-storage==2.13.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-auth==2.23.4
google-api-python-client==2.108.0

# AI/ML
vertexai==0.0.1
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2

# Text Processing
spacy==3.7.2
python-docx==1.0.1
PyPDF2==3.0.1
pdfminer.six==20221105

# Speech Processing
edge-tts==6.1.9
SpeechRecognition==3.10.0
pyaudio==0.2.13

# API Integrations
twilio==8.10.0
stripe==7.6.0
firebase-admin==6.2.0

# Email
aiosmtplib==2.0.2
jinja2==3.1.2
email-validator==2.1.0.post1

# Data Visualization
plotly==5.18.0
matplotlib==3.8.1
seaborn==0.13.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.5
passlib==1.7.4

# File Handling
python-magic==0.4.27
aiofiles==23.2.1
pillow==10.1.0

# Async Support
aiohttp==3.8.6
asyncio==3.4.3
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests-mock==1.11.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
validators==0.22.0
pydantic==2.4.2
rich==13.6.0

# Development Tools
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Documentation
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.4.7

# Monitoring & Logging
sentry-sdk==1.32.0
prometheus-client==0.18.0
python-json-logger==2.0.7

# Optional but Recommended
uvicorn==0.24.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4